+++ src/main/java/com/example/RealMatch/oauth/dto/CustomOAuth2User.java
@@ -1,4 +1,4 @@
-package com.example.RealMatch.global.oauth.dto;
+package com.example.RealMatch.oauth.dto;
 
 import java.util.Collection;
 import java.util.Collections;
