+++ src/main/java/com/example/RealMatch/oauth/service/CustomOAuth2UserService.java
@@ -1,4 +1,4 @@
-package com.example.RealMatch.global.oauth.service;
+package com.example.RealMatch.oauth.service;
 
 import java.util.Map;
 
@@ -9,13 +9,13 @@ import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.example.RealMatch.global.oauth.code.OAuthErrorCode;
-import com.example.RealMatch.global.oauth.dto.CustomOAuth2User;
-import com.example.RealMatch.global.oauth.dto.GoogleUserInfo;
-import com.example.RealMatch.global.oauth.dto.KakaoUserInfo;
-import com.example.RealMatch.global.oauth.dto.NaverUserInfo;
-import com.example.RealMatch.global.oauth.dto.OAuth2UserInfo;
-import com.example.RealMatch.global.oauth.exception.AuthException;
+import com.example.RealMatch.oauth.code.OAuthErrorCode;
+import com.example.RealMatch.oauth.dto.CustomOAuth2User;
+import com.example.RealMatch.oauth.dto.GoogleUserInfo;
+import com.example.RealMatch.oauth.dto.KakaoUserInfo;
+import com.example.RealMatch.oauth.dto.NaverUserInfo;
+import com.example.RealMatch.oauth.dto.OAuth2UserInfo;
+import com.example.RealMatch.oauth.exception.AuthException;
 import com.example.RealMatch.user.domain.entity.AuthenticationMethod;
 import com.example.RealMatch.user.domain.entity.User;
 import com.example.RealMatch.user.domain.entity.enums.Role;
