+++ src/main/java/com/example/RealMatch/oauth/swagger/AuthSwagger.java
@@ -1,12 +1,12 @@
-package com.example.RealMatch.global.presentation.swagger;
+package com.example.RealMatch.oauth.swagger;
 
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 
 import com.example.RealMatch.global.config.jwt.CustomUserDetails;
-import com.example.RealMatch.global.oauth.dto.OAuthTokenResponse;
-import com.example.RealMatch.global.oauth.dto.request.SignupCompleteRequest;
 import com.example.RealMatch.global.presentation.CustomResponse;
+import com.example.RealMatch.oauth.dto.OAuthTokenResponse;
+import com.example.RealMatch.oauth.dto.request.SignupCompleteRequest;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
